requirements.txt
click
requests
psycopg2-binary
pyyaml
